cleanup in Som::train

+++ include/SOM.hpp
@@ -50,7 +50,7 @@ public:
 	Som(size_t, size_t, size_t, bool verbose = false);
 	Som(const char *, bool verbose = false);
 	~Som() = default;
-	void train(DataSet *, int, double, double, double, double, int);
+	void train(DataSet *, size_t, double, double, double, double, int);
 	double evaluate(const DataSet *) const;
 	SomIndex trainSingle(const Eigen::VectorXf&, const std::vector<int>&, const std::vector<double>&, const double, const double, size_t &, const int);
 	int measureSimilarity(const DataSet *, int, int) const;
